Started to re-write the validations section
This is a re-write both for greater clarity, and to remove the redundant
mentions of the Validatable gem.  The intent is to present the reader first with
a simple case of validations, and then explain the more complex context specific
stuff, as well as updating that section to use the current terminology.